Directory structure refactoring. Add individual posts.
