+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.css';
-import Header from "./components/Header";
-import Navbar from "./components/Navbar";
-import Profile from "./components/Profile";
+import Header from "./components/header/Header";
+import Navbar from "./components/navbar/Navbar";
+import Profile from "./components/profile/Profile";
 
 const App = () => {
     return (
