+++ src/components/header/Header.jsx
@@ -6,7 +6,8 @@ const Header = () => {
     return (
     <header className={classes.header}>
         <img src='https://cdn.shopify.com/shopifycloud/hatchful-web/assets/6fcc76cfd1c59f44d43a485167fb3139.png'/>
-    </header>);
+    </header>
+    );
 };
 
 export default Header;
