+++ src/components/profile/Profile.module.css
@@ -1,0 +1,4 @@
+.content {
+    grid-area: c;
+    background-color: lightblue;
+}
