+++ src/components/profile/myPosts/MyPosts.module.css
@@ -1,0 +1,11 @@
+.content {
+    grid-area: c;
+    background-color: lightblue;
+}
+
+.item{
+    color: greenyellow;
+}
+.item img{
+    width: 50px;
+}
