+++ src/components/profile/myPosts/post/Post.module.css
@@ -6,3 +6,8 @@
 .item{
     color: greenyellow;
 }
+.item img{
+    height: 50px;
+    width: 50px;
+    border-radius: 50px;
+}
